job-alert-resume/requirements.txt
# Gmail
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0

# HTTP & parsing
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Indeed (JS-rendered pages)
playwright>=1.40.0

# Config
python-dotenv>=1.0.0

# Gemini AI (use one; code supports both)
google-genai>=1.0.0
google-generativeai>=0.8.0
# OpenAI (optional; set JAR_AI_PROVIDER=openai and OPENAI_API_KEY to use ChatGPT)
openai>=1.0.0
# Markdown → HTML for PDF export
markdown>=3.5.0
